refactor(e2e): name conversion[frontend]
