+++ frontend/e2e/fixtures/create-new/config-step.ts
