+++ frontend/e2e/fixtures/create-new/metric-data.csv
