+++ frontend/e2e/fixtures/create-new/metrics-step.ts
