+++ frontend/e2e/fixtures/cycle-time-by-status/board-data-by-status.csv
