+++ frontend/e2e/fixtures/cycle-time-by-status/cycle-time-by-column-fixture.ts
