+++ frontend/e2e/fixtures/cycle-time-by-status/cycle-time-by-status-fixture.ts
