+++ frontend/e2e/fixtures/cycle-time-by-status/metric-data-by-status.csv
