+++ frontend/e2e/fixtures/import-file/metric-data.csv
