+++ frontend/e2e/specs/major-path/create-a-new-project.spec.ts
@@ -1,11 +1,11 @@
-import { cycleTimeByStatusFixture } from '../../fixtures/cycleTimeByStatus/cycleTimeByStatusFixture';
-import { BOARD_METRICS_RESULT, DORA_METRICS_RESULT } from '../../fixtures/createNew/reportResult';
-import { config as metricsStepData } from '../../fixtures/createNew/metricsStep';
-import { config as configStepData } from '../../fixtures/createNew/configStep';
-import { ProjectCreationType } from '../../pages/metrics/ReportStep';
-import { test } from '../../fixtures/testWithExtendFixtures';
-import { clearTempDir } from '../../utils/clearTempDir';
-import { format } from '../../utils/dateTime';
+import { cycleTimeByStatusFixture } from '../../fixtures/cycle-time-by-status/cycle-time-by-status-fixture';
+import { BOARD_METRICS_RESULT, DORA_METRICS_RESULT } from '../../fixtures/create-new/report-result';
+import { config as metricsStepData } from '../../fixtures/create-new/metrics-step';
+import { config as configStepData } from '../../fixtures/create-new/config-step';
+import { ProjectCreationType } from 'e2e/pages/metrics/report-step';
+import { test } from '../../fixtures/test-with-extend-fixtures';
+import { clearTempDir } from 'e2e/utils/clear-temp-dir';
+import { format } from 'e2e/utils/date-time';
 
 test.beforeAll(async () => {
   await clearTempDir();
