+++ frontend/e2e/specs/major-path/page-jumps.spec.ts
@@ -1,11 +1,11 @@
 import {
   config as metricsStepData,
   modifiedConfig as modifiedMetricsStepData,
-} from '../../fixtures/createNew/metricsStep';
-import { config as configStepData } from '../../fixtures/createNew/configStep';
-import { test } from '../../fixtures/testWithExtendFixtures';
-import { clearTempDir } from '../../utils/clearTempDir';
-import { format } from '../../utils/dateTime';
+} from '../../fixtures/create-new/metrics-step';
+import { config as configStepData } from '../../fixtures/create-new/config-step';
+import { test } from '../../fixtures/test-with-extend-fixtures';
+import { clearTempDir } from 'e2e/utils/clear-temp-dir';
+import { format } from 'e2e/utils/date-time';
 
 test.beforeAll(async () => {
   await clearTempDir();
