+++ frontend/e2e/specs/side-path/unhappy-path.spec.ts
@@ -1,11 +1,11 @@
 import {
   importInputWrongProjectFromFile as importUnhappyPathProjectFromFile,
   importModifiedCorrectConfig as modifiedCorrectProjectFromFile,
-} from '../../fixtures/importFile/unhappy-path-file';
-import { BOARD_METRICS_RESULT, DORA_METRICS_RESULT } from '../../fixtures/createNew/reportResult';
-import { test } from '../../fixtures/testWithExtendFixtures';
-import { clearTempDir } from '../../utils/clearTempDir';
-import { format } from '../../utils/dateTime';
+} from '../../fixtures/import-file/unhappy-path-file';
+import { BOARD_METRICS_RESULT, DORA_METRICS_RESULT } from '../../fixtures/create-new/report-result';
+import { test } from '../../fixtures/test-with-extend-fixtures';
+import { clearTempDir } from '../../utils/clear-temp-dir';
+import { format } from '../../utils/date-time';
 
 test.beforeAll(async () => {
   await clearTempDir();
